fix small issue elorapports
