--- data/reeks_v2.csv
@@ -1,105 +1,0 @@
-Nr,Club1,Club2,Club3,Club4
-,Division 1 Afdeling,,Division 2A Afdeling,Division 2B Afdeling
-1,301 KOSK 1,,703 Eisden/MSK-Dilsen 1,301 KOSK 2
-2,471 Wachtebeke 1,,176 Rokade Westerlo 1,109 Borgerhout 1
-3,432 Wetteren 1,,401 KGSRL 2,114 Mechelen 1
-4,621 T.A.L 621 1,,952 Wavre 1,401 KGSRL 3
-5,166 TSM Mechelen 1,,239 Boitsfort 1,166 TSM Mechelen 2
-6,601 CRELEL 1,,226 Europchess 1,462 Zottegem 1
-7,303 KBSK Brugge 1,,501 CREC Charleroi 1,303 KBSK Brugge 2
-8,401 KGSRL 1,,618 Echiquier Mosan 1,278 Pantin 1
-9,402 Jean Jaurès Gent 1,,607 KSK Rochade 2,261 Gambiet Opwijk 1
-10,607 KSK Rochade 1,,514 MCC 1,209 The Belgian CC 1
-11,230 Leuven Centraal 1,,231 Dolle Toren Leuven 1,436 LSV-Chesspirant 1
-12,901 Namur Echecs 1,,548 Caïssa Europe 1,124 Deurne 1
-,Division 3A Afdeling,Division 3B Afdeling,Division 3C Afdeling,Division 3D Afdeling
-1,174 Brasschaat 1,207 2 Fous Diogène 1,302 KISK Ieper 1,727 Midden Limburg 1
-2,101 KASK 1,244 Brussels 1,340 Izegem 1,244 Brussels 2
-3,114 Mechelen 2,228 Dworp 1,432 Wetteren 2,627 Wirtzfeld 1
-4,260 Kapelle-op-den Bos 1,239 Boitsfort 2,309 KRST 1,810 Marche-en-Famenne 1
-5,401 KGSRL 4,417 Pion-Aalst 1,313 KWSLE 1,604 KSK47-Eynatten 1
-6,128 Beveren 1,226 Europchess 2,541 Leuze 1,601 CRELEL 2
-7,132 Oude God 1,548 Caïssa Europe 2,303 KBSK Brugge 3,601 CRELEL 3
-8,410 St.-Niklaas 1,278 Pantin 2,472 De Mercatel 1,230 Leuven Centraal 2
-9,708 NLS Lommel 1,201 CREB Bruxelles 1,521 Tournai 1,607 KSK Rochade 3
-10,143 Boey Temse 1,401 KGSRL 5,401 KGSRL 6,135 Geel 1
-11,425 Dendermonde 1,511 EDC 1,436 LSV-Chesspirant 2,231 Dolle Toren Leuven 2
-12,174 Brasschaat 2,418 VPTD Geraardsbergen 1,430 Landegem 1,901 Namur Echecs 2
-,Division 4A Afdeling,Division 4B Afdeling,Division 4C Afdeling,Division 4D Afdeling
-1,727 Midden Limburg 2,174 Brasschaat 3,301 KOSK 3,901 Namur Echecs 3
-2,712 Landen 1,109 Borgerhout 2,302 KISK Ieper 2,230 Leuven Centraal 3
-3,627 Wirtzfeld 2,101 KASK 2,432 Wetteren 3,228 Dworp 2
-4,621 T.A.L 621 2,436 LSV-Chesspirant 3,422 MSV 1,909 Philippeville 1
-5,604 KSK47-Eynatten 2,714 Overpelt 1,521 Tournai 2,551 HCC Jurbise 1
-6,601 CRELEL 4,194 ChessLooks Lier 1,307 Bredene 1,601 CRELEL 6
-7,601 CRELEL 5,132 Oude God 2,303 KBSK Brugge 4,501 CREC Charleroi 2
-8,618 Echiquier Mosan 2,465 Artevelde 1,472 De Mercatel 2,952 Wavre 2
-9,607 KSK Rochade 4,130 Moretus Hoboken 1,438 Deinze 1,607 KSK Rochade 6
-10,607 KSK Rochade 5,135 Geel 2,304 Tielt 1,547 Renaissance Binche 1
-11,121 TUSK 1,121 TUSK 2,401 KGSRL 7,902 C.E.Sambrevillois 1
-12,622 Herve 1,124 Deurne 2,430 Landegem 2,525 Anderlues 1
-,Division 4E Afdeling,Division 4F Afdeling,Division 4G Afdeling,Division 4H Afdeling
-1,207 2 Fous Diogène 2,207 2 Fous Diogène 3,401 KGSRL 9,301 KOSK 4
-2,230 Leuven Centraal 4,244 Brussels 3,471 Wachtebeke 2,471 Wachtebeke 3
-3,432 Wetteren 4,114 Mechelen 3,109 Borgerhout 3,432 Wetteren 5
-4,233 DZD Halle 1,230 Leuven Centraal 5,436 LSV-Chesspirant 4,436 LSV-Chesspirant 5
-5,401 KGSRL 8,229 Woluwé 1,166 TSM Mechelen 3,417 Pion-Aalst 2
-6,226 Europchess 3,601 CRELEL 7,194 ChessLooks Lier 2,462 Zottegem 2
-7,902 C.E.Sambrevillois 2,226 Europchess 4,132 Oude God 3,303 KBSK Brugge 5
-8,278 Pantin 3,278 Pantin 4,465 Artevelde 2,472 De Mercatel 3
-9,261 Gambiet Opwijk 2,201 CREB Bruxelles 2,130 Moretus Hoboken 2,402 Jean Jaurès Gent 2
-10,209 The Belgian CC 2,209 The Belgian CC 3,230 Leuven Centraal 6,460 Oudenaarde 1
-11,961 Braine Echecs 1,231 Dolle Toren Leuven 3,121 TUSK 3,401 KGSRL 10
-12,525 Anderlues 2,952 Wavre 3,124 Deurne 3,430 Landegem 3
-,Division 5A Afdeling,Division 5B Afdeling,Division 5C Afdeling,Division 5D Afdeling
-1,901 Namur Echecs 4,901 Namur Echecs 5,623 Sankt Vith 1,301 KOSK 5
-2,244 Brussels 4,548 Caïssa Europe 3,811 Arlon 1,340 Izegem 2
-3,114 Mechelen 4,902 C.E.Sambrevillois 3,627 Wirtzfeld 3,422 MSV 2
-4,239 Boitsfort 3,909 Philippeville 2,816 Bastogne 1,309 KRST 2
-5,230 Leuven Centraal 7,551 HCC Jurbise 2,604 KSK47-Eynatten 3,313 KWSLE 2
-6,261 Gambiet Opwijk 3,549 Saint-Ghislain 1,601 CRELEL 8,307 Bredene 2
-7,226 Europchess 5,501 CREC Charleroi 3,703 Eisden/MSK-Dilsen 2,303 KBSK Brugge 6
-8,278 Pantin 5,000 BYE 5B,618 Echiquier Mosan 3,475 Rapid Aalter 1
-9,201 CREB Bruxelles 3,518 Soignies 1,607 KSK Rochade 7,322 KVSK Veurne 1
-10,952 Wavre 4,514 MCC 2,609 Anthisnes 1,351 Knokke Schaak 1
-11,961 Braine Echecs 2,953 Nivelles 1,619 Welkenraedt 1,401 KGSRL 11
-12,207 2 Fous Diogène 4,525 Anderlues 3,000 BYE 5C,000 BYE 5D
-,Division 5E Afdeling,Division 5F Afdeling,Division 5G Afdeling,Division 5H Afdeling
-1,727 Midden Limburg 3,727 Midden Limburg 4,174 Brasschaat 4,302 KISK Ieper 3
-2,176 Rokade Westerlo 2,176 Rokade Westerlo 3,176 Rokade Westerlo 4,340 Izegem 3
-3,621 T.A.L 621 3,114 Mechelen 5,114 Mechelen 6,432 Wetteren 6
-4,162 Mol 1,162 Mol 2,162 Mol 3,309 KRST 3
-5,714 Overpelt 2,166 TSM Mechelen 4,194 ChessLooks Lier 3,313 KWSLE 3
-6,601 CRELEL 9,261 Gambiet Opwijk 4,128 Beveren 2,401 KGSRL 13
-7,703 Eisden/MSK-Dilsen 3,132 Oude God 4,132 Oude God 5,303 KBSK Brugge 7
-8,736 Heusden 1,278 Pantin 6,401 KGSRL 12,475 Rapid Aalter 2
-9,000 BYE 5E,230 Leuven Centraal 8,130 Moretus Hoboken 3,438 Deinze 2
-10,135 Geel 3,260 Kapelle-op-den Bos 2,230 Leuven Centraal 9,304 Tielt 2
-11,121 TUSK 4,231 Dolle Toren Leuven 4,195 Chessmates  1,436 LSV-Chesspirant 6
-12,182 Noorderwijk 1,124 Deurne 4,124 Deurne 5,000 BYE 5H
-,Division 5I Afdeling,Division 5J Afdeling,Division 5K Afdeling,Division 5L Afdeling
-1,207 2 Fous Diogène 5,000 BYE 5J,174 Brasschaat 5,727 Midden Limburg 5
-2,244 Brussels 5,471 Wachtebeke 4,401 KGSRL 16,712 Landen 2
-3,228 Dworp 3,114 Mechelen 7,114 Mechelen 8,621 T.A.L 621 4
-4,233 DZD Halle 2,239 Boitsfort 4,196 De Rode Loper 1,609 Anthisnes 2
-5,401 KGSRL 14,417 Pion-Aalst 3,166 TSM Mechelen 5,230 Leuven Centraal 12
-6,541 Leuze 2,462 Zottegem 3,128 Beveren 3,601 CRELEL 10
-7,548 Caïssa Europe 4,230 Leuven Centraal 10,132 Oude God 6,666 666 1
-8,278 Pantin 7,278 Pantin 8,230 Leuven Centraal 11,618 Echiquier Mosan 4
-9,201 CREB Bruxelles 4,261 Gambiet Opwijk 5,130 Moretus Hoboken 4,703 Eisden/MSK-Dilsen 4
-10,514 MCC 3,209 The Belgian CC 4,143 Boey Temse 2,607 KSK Rochade 8
-11,961 Braine Echecs 3,436 LSV-Chesspirant 7,195 Chessmates  2,962 CCBBE 1
-12,000 BYE 5I,401 KGSRL 15,190 Burcht 1,733 St.-Truiden 1
-,Division 5M Afdeling,Division 5N Afdeling,Division 5O Afdeling,Division 5P Afdeling
-1,401 KGSRL 17,174 Brasschaat 6,207 2 Fous Diogène 6,901 Namur Echecs 6
-2,471 Wachtebeke 5,109 Borgerhout 4,176 Rokade Westerlo 5,230 Leuven Centraal 14
-3,432 Wetteren 7,101 KASK 3,114 Mechelen 8,521 Tournai 3
-4,422 MSV 3,436 LSV-Chesspirant 9,239 Boitsfort 5,909 Philippeville 3
-5,313 KWSLE 4,401 KGSRL 18,229 Woluwé 2,551 HCC Jurbise 3
-6,462 Zottegem 4,128 Beveren 4,261 Gambiet Opwijk 6,541 Leuze 3
-7,465 Artevelde 3,132 Oude God 7,204 Excelsior 1,501 CREC Charleroi 4
-8,472 De Mercatel 4,410 St.-Niklaas 2,278 Pantin 9,548 Caïssa Europe 5
-9,402 Jean Jaurès Gent 3,130 Moretus Hoboken 5,401 KGSRL 19,902 C.E.Sambrevillois 4
-10,460 Oudenaarde 2,143 Boey Temse 3,260 Kapelle-op-den Bos 3,514 MCC 4
-11,436 LSV-Chesspirant 8,425 Dendermonde 2,231 Dolle Toren Leuven 5,953 Nivelles 2
-12,404 Drie Torens 1,190 Burcht 2,230 Leuven Centraal 13,549 Saint-Ghislain 2
