--- data/test_clubs.csv
@@ -1,16 +1,0 @@
-idclub,name,nplayers
-1001,FC Antwerp,59
-1002,Union,45
-1003,RC Genk,70
-1004,AA Gent,112
-1005,Club Brugge,50
-1006,KV Mechelen,40
-1007,KV Oostende,36
-1008,St Truiden,31
-1009,KV Kortijk,39
-1010,Standard,18
-1011,Zulte Waregem,22
-1012,RWDM.41
-1013,OHLeuven,7
-1014,Cercle Brugge,12
-1015,Westerlo,11
