--- data/test_series.csv
@@ -1,37 +1,0 @@
-div,series,pairing,idclub,name,team
-1,,1,1001,FC Antwerp,1
-1,,2,1002,Union,1
-1,,3,1003,RC Genk,1
-1,,4,1004,AA Gent,1
-1,,5,1005,Club Brugge,1
-1,,6,1006,KV Mechelen,1
-2,A,1,1005,Club Brugge,2
-2,A,2,1001,FC Antwerp,2
-2,A,3,1004,AA Gent,2
-2,A,4,1007,KV Oostende,1
-2,A,5,1008,St Truiden,1
-2,A,6,1009,KV Kortijk,1
-2,B,1,1002,Union,2
-2,B,2,1010,Standard,1
-2,B,3,1004,AA Gent,3
-2,B,4,1006,KV Mechelen,2
-2,B,5,1011,Zulte Waregem,1
-2,B,6,1012,RWDM,1
-3,A,1,1003,RC Genk,2
-3,A,2,1007,KV Oostende,2
-3,A,3,1001,FC Antwerp,3
-3,A,4,1001,FC Antwerp,4
-3,A,5,1008,St Truiden,2
-3,A,6,1008,St Truiden,3
-3,B,1,1002,Union,3
-3,B,2,1004,AA Gent,4
-3,B,3,1003,RC Genk,3
-3,B,4,1006,KV Mechelen,3
-3,B,5,1004,AA Gent,5
-3,B,6,1012,RWDM,2
-3,C,1,1004,AA Gent,6
-3,C,2,1011,Zulte Waregem,2
-3,C,3,1002,Union,4
-3,C,4,1007,KV Oostende,3
-3,C,5,1009,KV Kortijk,2
-3,C,6,1012,RWDM,3
