+++ kbsb/eloprocessing/fide_elo.py
@@ -165,84 +165,6 @@ async def games_round(round):
                 )
 
 
-def to_elo_players():
-    global sortedplayers
-    for g in games:
-        wt = tlines.setdefault(unidecode(g.team_white), [])
-        bt = tlines.setdefault(unidecode(g.team_black), [])
-        wopp = ""
-        bopp = ""
-        if g.result == "1-0":
-            wsc1 = 1.0
-            bsc1 = 0.0
-            wsc2 = "1"
-            bsc2 = "0"
-        if g.result == "½-½":
-            wsc1 = 0.5
-            bsc1 = 0.5
-            wsc2 = "="
-            bsc2 = "="
-        if g.result == "0-1":
-            wsc1 = 0.0
-            bsc1 = 1.0
-            wsc2 = "0"
-            bsc2 = "1"
-        if g.result == "1-0 FF":
-            wsc1 = 1.0
-            wsc2 = "+"
-            wopp = "0000"
-            bsc1 = 0.0
-            bsc2 = "-"
-            bopp = "0000"
-        if g.result == "0-1 FF":
-            wsc1 = 0.0
-            wsc2 = "-"
-            wopp = "0000"
-            bsc1 = 1.0
-            bsc2 = "+"
-            bopp = "0000"
-        white = EloPlayer(
-            idbel=g.idbel_white,
-            idfide=g.idfide_white,
-            fullname=g.fullname_white,
-            fiderating=g.fiderating_white,
-            natfide=g.natfide_white,
-            birthday=g.birthday_white,
-            title=g.title_white,
-            gender=g.gender_white,
-            sc1=wsc1,
-            sc2=wsc2,
-            idopp=g.idbel_black,
-            team=unidecode(g.team_white),
-            color="w",
-        )
-        black = EloPlayer(
-            idbel=g.idbel_black,
-            idfide=g.idfide_black,
-            fullname=g.fullname_black,
-            fiderating=g.fiderating_black,
-            natfide=g.natfide_black,
-            birthday=g.birthday_black,
-            title=g.title_black,
-            gender=g.gender_black,
-            sc1=bsc1,
-            sc2=bsc2,
-            idopp=g.idbel_white,
-            team=unidecode(g.team_black),
-            color="b",
-        )
-        elopl[white.idbel] = white
-        elopl[black.idbel] = black
-    sortedplayers = sorted(
-        elopl.keys(), key=lambda x: (-elopl[x].fiderating, elopl[x].fullname)
-    )
-    logger.info(f"sortedplayers {len(sortedplayers)}")
-    for ix, key in enumerate(sortedplayers):
-        elopl[key].myix = ix + 1
-        elopl[elopl[key].idopp].oppix = ix + 1
-        tlines[elopl[key].team].append(ix + 1)
-
-
 def to_fide_elo(round):
     global sortedplayers
     """
